requirements.txt
Flask==2.2.5
selenium==4.8.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
pytest==7.4.0
gunicorn==21.2.0
waitress==2.1.2
lxml==4.9.3
flask-wtf==1.2.1
python-dotenv==1.0.0
flask-cors==3.0.10
flask-socketio==5.3.4




